Сохранение игры против человека, осталось пара багов

+++ src/main/java/com.pesikovlike/kalah/server/websocket/game/gameWS.java
@@ -13,6 +13,7 @@ import com.pesikovlike.kalah.model.dao.AvatarDAO;
 import com.pesikovlike.kalah.model.dao.UserDAO;
 import com.pesikovlike.kalah.model.entity.Avatar;
 import com.pesikovlike.kalah.model.entity.GameState;
+import com.pesikovlike.kalah.model.entity.Hole;
 import com.pesikovlike.kalah.model.entity.User;
 import com.pesikovlike.kalah.user.UserService;
 
@@ -27,6 +28,8 @@ import javax.websocket.server.ServerEndpoint;
 import java.io.IOException;
 import java.io.StringReader;
 import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -60,6 +63,7 @@ public class gameWS {
 
     private GameSession gameSession;
     private String creatorLogin;
+    private String login;
     private GameBid gameBid;
     private AI ai;
 
@@ -67,11 +71,179 @@ public class gameWS {
 
     @OnMessage
     public void onMessage(String message, Session session) throws IOException, EncodeException {
-
+        LOGGER.log(Level.SEVERE, "Message: " + message);
         JsonParser parser = new JsonParser();
         com.google.gson.JsonObject mesg = parser.parse(message).getAsJsonObject();
         String operation = mesg.get("operation").getAsString();
-        LOGGER.log(Level.SEVERE, "Message: " + message);
+
+        if (operation.equals("load")) {
+            LOGGER.log(Level.SEVERE, "Start load");
+            login = mesg.get("login").getAsString();
+            creatorLogin = mesg.get("creatorLogin").getAsString();
+            long id = mesg.get("id").getAsLong();
+            LOGGER.log(Level.SEVERE, "Id: " + id);
+            GameSession ses = gameSessionService.getGameSessionForLoad(id);
+            
+            if (ses == null) {
+                LOGGER.log(Level.SEVERE, "ses == null");
+                List<GameState> games = userService.getSavedGamesForUser(login);
+                Iterator<GameState> iter = games.iterator();
+                GameState state = null;
+                while (iter.hasNext()) {
+                    state = iter.next();
+                    if (state.getGameStateId() == id) {
+                        break;
+                    }
+                }
+                if (creatorLogin.equals(login)) {
+                    LOGGER.log(Level.SEVERE, "It's creator");
+                    gameSession = gameSessionService.addGameSession(state, session, null);
+                } else {
+                    LOGGER.log(Level.SEVERE, "It's joined");
+                    gameSession = gameSessionService.addGameSession(state, null, session);
+                }
+                LOGGER.log(Level.SEVERE, "After add");
+
+                Map<String, String> resultMap;
+                Gson gson = new GsonBuilder().setPrettyPrinting().create();
+                String json;
+                String role = mesg.get("role").getAsString();
+                resultMap = new HashMap<String, String>();
+                resultMap.put("operation", "load");
+                resultMap.put("suboperation", "create");
+                resultMap.put("role", role);
+                int numHoles = gameSession.getGameState().getInitialHoleCount();
+                resultMap.put("holeCount", String.valueOf(numHoles));
+                resultMap.put("stoneCount", String.valueOf(gameSession.getGameState().getInitialStoneCount()));
+                Hole[] holes = gameSession.getHolesArray();
+
+                for (int i = 0; i < holes.length; i++) {
+                    resultMap.put(String.valueOf(holes[i].getNumber()), String.valueOf(holes[i].getStoneCount()));
+                }
+                if (creatorLogin.equals(login)) {
+                    if (gameSession.getGameState().getPriority()) {
+                        resultMap.put("priority", "true");
+                    } else {
+                        resultMap.put("priority", "false");
+                    }
+                } else {
+                    if (gameSession.getGameState().getPriority()) {
+                        resultMap.put("priority", "false");
+                    } else {
+                        resultMap.put("priority", "true");
+                    }
+                }
+
+                User user = userDAO.getUserByLogin(mesg.get("login").getAsString());
+                Avatar avatar = avatarDAO.getAvatarById(user.getAvatar().getAvatarId());
+                resultMap.put("yourAvatar", avatar.getFilePath());
+                resultMap.put("login", login);
+
+                if (creatorLogin.equals(login)) {
+                    user = userDAO.getUserByLogin(gameSession.getGameState().getUser2().getLogin());
+                } else {
+                    user = userDAO.getUserByLogin(gameSession.getGameState().getUser1().getLogin());
+                }
+                avatar = avatarDAO.getAvatarById(user.getAvatar().getAvatarId());
+                resultMap.put("enemyLogin", user.getLogin());
+                resultMap.put("enemyAvatar", avatar.getFilePath());
+
+                json = gson.toJson(resultMap);
+                LOGGER.log(Level.SEVERE, "response: " + json);
+                session.getBasicRemote().sendText(json);
+
+
+            } else {
+                LOGGER.log(Level.SEVERE, "ses != null");
+                List<GameState> games = userService.getSavedGamesForUser(login);
+                Iterator<GameState> iter = games.iterator();
+                GameState state = null;
+                while (iter.hasNext()) {
+                    state = iter.next();
+                    if (state.getGameStateId() == id) {
+                        break;
+                    }
+                }
+                if (creatorLogin.equals(login)) {
+                    LOGGER.log(Level.SEVERE, "It's creator");
+                    gameSession = gameSessionService.getGameSession(creatorLogin);
+                    gameSession.setSessionOfCreator(session);
+                } else {
+                    LOGGER.log(Level.SEVERE, "It's joined");
+                    gameSession = gameSessionService.getGameSession(creatorLogin);
+                    gameSession.setSessionOfJoined(session);
+                }
+                LOGGER.log(Level.SEVERE, "After add");
+
+                Map<String, String> resultMap;
+                Gson gson = new GsonBuilder().setPrettyPrinting().create();
+                String json;
+                String role = mesg.get("role").getAsString();
+                resultMap = new HashMap<String, String>();
+                resultMap.put("operation", "load");
+                resultMap.put("suboperation", "join");
+                resultMap.put("role", role);
+
+                int numHoles = gameSession.getGameState().getInitialHoleCount();
+                resultMap.put("holeCount", String.valueOf(numHoles));
+                resultMap.put("stoneCount", String.valueOf(gameSession.getGameState().getInitialStoneCount()));
+                Hole[] holes = gameSession.getHolesArray();
+
+                for (int i = 0; i < holes.length; i++) {
+                    resultMap.put(String.valueOf(holes[i].getNumber()), String.valueOf(holes[i].getStoneCount()));
+                }
+                String priority = "";
+                if (creatorLogin.equals(login)) {
+                    if (gameSession.getGameState().getPriority()) {
+                        priority = "true";
+                    } else {
+                        priority = "false";
+                    }
+                } else {
+                    if (gameSession.getGameState().getPriority()) {
+                        priority = "false";
+                    } else {
+                        priority = "true";
+                    }
+                }
+                resultMap.put("priority", priority);
+                User user = userDAO.getUserByLogin(mesg.get("login").getAsString());
+                Avatar avatar = avatarDAO.getAvatarById(user.getAvatar().getAvatarId());
+                resultMap.put("yourAvatar", avatar.getFilePath());
+                resultMap.put("login", login);
+
+                if (creatorLogin.equals(login)) {
+                    user = userDAO.getUserByLogin(gameSession.getGameState().getUser2().getLogin());
+                } else {
+                    user = userDAO.getUserByLogin(gameSession.getGameState().getUser1().getLogin());
+                }
+                avatar = avatarDAO.getAvatarById(user.getAvatar().getAvatarId());
+                resultMap.put("enemyLogin", user.getLogin());
+                resultMap.put("enemyAvatar", avatar.getFilePath());
+
+                json = gson.toJson(resultMap);
+                LOGGER.log(Level.SEVERE, "response: " + json);
+                session.getBasicRemote().sendText(json);
+
+                resultMap = new HashMap<String, String>();
+                resultMap.put("operation", "continue");
+                if (priority.equals("true")) {
+                    resultMap.put("priority", "false");
+                } else {
+                    resultMap.put("priority", "true");
+                }
+
+                json = gson.toJson(resultMap);
+                LOGGER.log(Level.SEVERE, "response: " + json);
+                if (creatorLogin.equals(login)) {
+                    gameSession.getSessionOfJoined().getBasicRemote().sendText(json);
+                } else {
+                    gameSession.getSessionOfCreator().getBasicRemote().sendText(json);
+                }
+
+            }
+            return;
+        }
 
         if (operation.equals("createAI")) {
             String level = mesg.get("level").getAsString();
@@ -136,6 +308,7 @@ public class gameWS {
                     gameSession = gameSessionService.addGameSession(gameBid);
                     gameSession.getGameState().setUser2(userDAO.getUserByLogin(joinedLogin));
                     gameBidService.deleteBid(gameBid.getCreatorLogin());
+                    gameBid = null;
                     LOGGER.log(Level.SEVERE, "Start yes");
                     Map<String, String> resultMap;
                     Gson gson = new GsonBuilder().setPrettyPrinting().create();
@@ -271,6 +444,7 @@ public class gameWS {
 
             session.getBasicRemote().sendText(json);
         }
+
     }
 
     //
@@ -286,16 +460,17 @@ public class gameWS {
         String json;
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
-        if (gameBid.getSessionOfCreator().equals(session)) {
+        if (gameBid != null && gameBid.getSessionOfCreator().equals(session)) {
             if (gameBid.getSessionOfJoined() != null) {
                 resultMap = new HashMap<String, String>();
                 resultMap.put("operation", "creator closed");
 
                 json = gson.toJson(resultMap);
                 gameBid.getSessionOfJoined().getBasicRemote().sendText(json);
+                gameBid = null;
             }
             gameBidService.deleteBid(creatorLogin);
-        } else if (gameBid.getSessionOfJoined().equals(session)) {
+        } else if (gameBid != null && gameBid.getSessionOfJoined().equals(session)) {
             gameBid.setSessionOfJoined(null);
             resultMap = new HashMap<String, String>();
             resultMap.put("operation", "joined closed");
@@ -303,22 +478,33 @@ public class gameWS {
             json = gson.toJson(resultMap);
             gameBid.getSessionOfCreator().getBasicRemote().sendText(json);
             gameBid.setBlock(false);
+            gameBid = null;
         } else {
-            if (gameSession.getSessionOfCreator().equals(session)) {
+            LOGGER.log(Level.SEVERE, "leave game session");
+            if (gameSession.getSessionOfCreator() != null && gameSession.getSessionOfCreator().equals(session)) {
+                LOGGER.log(Level.SEVERE, "creator");
                 resultMap = new HashMap<String, String>();
                 resultMap.put("operation", "creator closed in game");
 
                 json = gson.toJson(resultMap);
-                gameBid.getSessionOfJoined().getBasicRemote().sendText(json);
-            } else if (gameSession.getSessionOfJoined().equals(session)) {
+                LOGGER.log(Level.SEVERE, "creator");
+                if (gameSession.getSessionOfJoined() != null) {
+                    LOGGER.log(Level.SEVERE, "creator");
+                    gameSession.getSessionOfJoined().getBasicRemote().sendText(json);
+                }
+            } else if (gameSession.getSessionOfJoined() != null && gameSession.getSessionOfJoined().equals(session)) {
+                LOGGER.log(Level.SEVERE, "joined");
                 resultMap = new HashMap<String, String>();
                 resultMap.put("operation", "joined closed in game");
 
                 json = gson.toJson(resultMap);
-                gameBid.getSessionOfJoined().getBasicRemote().sendText(json);
+                if (gameSession.getSessionOfCreator() != null)
+                    gameSession.getSessionOfCreator().getBasicRemote().sendText(json);
             }
             //TODO тут надо сохранить игру
             gameSessionService.deleteGameSession(creatorLogin);
+            gameSession = null;
+            LOGGER.log(Level.SEVERE, "gs count: " + gameSessionService.getAllGameSessions().size());
         }
     }
 }
